+++ pcm1datastitcher.h
@@ -104,7 +104,7 @@ private:
     uint16_t getFirstFieldLineNum(uint8_t in_order);
     uint16_t getSecondFieldLineNum(uint8_t in_order);
     uint16_t addLinesFromField(std::vector<PCM1SubLine> *field_buf, uint16_t ind_start, uint16_t count, uint16_t *last_line_num = NULL);
-    uint16_t addFieldPadding(uint16_t in_frame, uint16_t line_cnt, uint16_t *last_line_num = NULL);
+    uint16_t addFieldPadding(uint32_t in_frame, uint16_t line_cnt, uint16_t *last_line_num = NULL);
     void fillFirstFieldForOutput();
     void fillSecondFieldForOutput();
     void setBlockSampleRate(PCM1DataBlock *in_block);
